demo(android): unify collision + view-node into picking-collision (#2239 Batch 1)
Consolidates the retired `collision` and `view-node` demos into a single
`picking-collision` entry with a top segmented-button toggle:

- **Ray Hit-Test** (default) — tap-driven `CollisionSystem` highlights on
  five primitives (3 cubes, 2 spheres) + Reset Colors on-screen action.
- **View Node** — embedding live Compose UI as a textured quad inside a
  spinning 3D parent node, with hoisted tap state shared between an
  embedded Button and the viewport gesture listener.

Each sub-mode keeps its own `SceneView` (independent camera framing +
gesture pipeline) — pure UI consolidation, no logic rewrite.

### Deep-link aliases

`collision` and `view-node` redirect to `picking-collision` through
`DeepLinkRouter.DEMO_ID_ALIASES`. The unified demo opens on its default
Ray Hit-Test tab; tab pre-selection from alias is a follow-up.

`.maestro/android/interaction.yaml` keeps referring to `collision`,
`view-node` — they continue to pass through the alias since the flow only
asserts the per-demo crash gate (`Navigate back` visible).

### Tests

- `DemoInteractionTest.collision_shapeTapAndReset` — opens unified demo
  on default tab, exercises shape taps + Reset Colors action.
- `DemoInteractionTest.viewNode_visibleAndTapCounter` — opens unified
  demo, taps "View Node" segmented button, exercises viewport tap counter
  + Visible toggle.
- `DemoRenderingScreenshotTest.pickingCollisionDemo_default_state` —
  renamed from `collisionDemo_default_state`, points at the new
  `picking-collision` slug.
- Stale `collision_default.png` and `viewnode_default.png` goldens
  removed; harness will re-capture on first run.

### Quality gates

- ✅ `./gradlew :samples:android-demo:assembleDebug`
- ✅ `./gradlew :sceneview:detekt :arsceneview:detekt :sceneview-core:detekt`
- ✅ `bash samples/android-demo/scripts/collate-demos.sh` (60 fragments)

### Out of scope

- iOS mirror — Android-only per Batch 1 plan.
- Tab pre-selection from alias — follow-up.

Refs #2239 Batch 1.
