+++ llms.txt
@@ -4016,9 +4016,8 @@ by `samples/android-demo/scripts/collate-demos.sh` — never edit between the ma
 ### Interaction
 
 - `camera-controls` — Camera Controls. Orbit, pan, and free camera modes.
-- `collision` — Collision & Hit Test. Hit testing and collision detection.
 - `gesture-editing` — Gesture Editing. Move, scale, and rotate with gestures.
-- `view-node` — ViewNode. Native UI embedded in 3D space.
+- `picking-collision` — Picking & Collision. Ray hit-test and interactive ViewNode overlays.
 
 ### Advanced
