+++ samples/android-demo/src/debug/java/io/github/sceneview/demo/DemoHostActivity.kt
@@ -21,7 +21,6 @@ import io.github.sceneview.demo.demos.ARTerrainAnchorDemo
 import io.github.sceneview.demo.demos.AnimationDemo
 import io.github.sceneview.demo.demos.BillboardDemo
 import io.github.sceneview.demo.demos.CameraControlsDemo
-import io.github.sceneview.demo.demos.CollisionDemo
 import io.github.sceneview.demo.demos.CustomMeshDemo
 import io.github.sceneview.demo.demos.DebugOverlayDemo
 import io.github.sceneview.demo.demos.DynamicSkyDemo
@@ -35,13 +34,13 @@ import io.github.sceneview.demo.demos.LinesPathsDemo
 import io.github.sceneview.demo.demos.ModelViewerDemo
 import io.github.sceneview.demo.demos.MultiModelDemo
 import io.github.sceneview.demo.demos.PhysicsDemo
+import io.github.sceneview.demo.demos.PickingAndCollisionDemo
 import io.github.sceneview.demo.demos.PostProcessingDemo
 import io.github.sceneview.demo.demos.ReflectionProbesDemo
 import io.github.sceneview.demo.demos.SecondaryCameraDemo
 import io.github.sceneview.demo.demos.ShapeDemo
 import io.github.sceneview.demo.demos.TextDemo
 import io.github.sceneview.demo.demos.VideoDemo
-import io.github.sceneview.demo.demos.ViewNodeDemo
 import io.github.sceneview.demo.theme.SceneViewDemoTheme
 
 /**
@@ -119,8 +118,8 @@ class DemoHostActivity : ComponentActivity() {
             // Interaction
             "camera-controls" -> CameraControlsDemo(onBack = back)
             "gesture-editing" -> GestureEditingDemo(onBack = back)
-            "collision" -> CollisionDemo(onBack = back)
-            "view-node" -> ViewNodeDemo(onBack = back)
+            // #2239 Batch 1 — `collision` and `view-node` consolidated into `picking-collision`.
+            "picking-collision", "collision", "view-node" -> PickingAndCollisionDemo(onBack = back)
             // Advanced
             "physics" -> PhysicsDemo(onBack = back)
             "post-processing" -> PostProcessingDemo(onBack = back)
