--- samples/android-demo/src/main/res/values/strings_demo_collision.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `collision` demo (registered by
-     `CollisionFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_collision_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_collision_title">Collision &amp; Hit Test</string>
-    <string name="demo_collision_subtitle">Hit testing and collision detection</string>
-</resources>
