+++ samples/android-demo/src/main/res/values/strings_demo_picking_collision.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Strings for the unified `picking-collision` demo (registered by
+     `PickingAndCollisionFragment.kt`). Consolidates the retired `collision`
+     and `view-node` demos (#2239 Batch 1). Old ids continue to deep-link
+     through `DeepLinkRouter.DEMO_ID_ALIASES`. -->
+<resources>
+    <string name="demo_picking_collision_title">Picking &amp; Collision</string>
+    <string name="demo_picking_collision_subtitle">Ray hit-test and interactive ViewNode overlays</string>
+</resources>
