--- samples/android-demo/src/main/res/values/strings_demo_view_node.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `view-node` demo (registered by
-     `ViewNodeFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_view_node_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_view_node_title">ViewNode</string>
-    <string name="demo_view_node_subtitle">Native UI embedded in 3D space</string>
-</resources>
